Refactor Aspire.Shared project structure
Reintroduced the `Aspire.Shared` project with a new structure:
- Added `Aspire.Shared.csproj` with dependency injection support.
- Introduced `ProductDto.cs` for product-related DTOs.
- Added `ServiceCollectionExtensions.cs` for shared DI extensions.
- Created `Product.cs` to define the `Product` entity model.

These changes centralize shared functionality for better modularity.
