+++ VS2026DevProductivity.sln
@@ -30,6 +30,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspireApp.AppHost", "src\As
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspireApp.ServiceDefaults", "src\AspireApp.ServiceDefaults\AspireApp.ServiceDefaults.csproj", "{C2307AD8-48E1-32C0-70AC-DAC131D3041F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspireApp.SharedLib", "src\AspireApp.Shared\AspireApp.SharedLib.csproj", "{F806201F-2208-4E97-BBD1-1F2B3C527B3D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,6 +46,10 @@ Global
 		{C2307AD8-48E1-32C0-70AC-DAC131D3041F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C2307AD8-48E1-32C0-70AC-DAC131D3041F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C2307AD8-48E1-32C0-70AC-DAC131D3041F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F806201F-2208-4E97-BBD1-1F2B3C527B3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F806201F-2208-4E97-BBD1-1F2B3C527B3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F806201F-2208-4E97-BBD1-1F2B3C527B3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F806201F-2208-4E97-BBD1-1F2B3C527B3D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -52,6 +58,7 @@ Global
 		{B96C39A2-528A-4AFF-95AE-76CB7239E6AB} = {4A1974BF-69D5-4714-B5E7-93BEFF8FC2AC}
 		{F8394078-E883-4F56-AA36-AB0D803A539D} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{C2307AD8-48E1-32C0-70AC-DAC131D3041F} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{F806201F-2208-4E97-BBD1-1F2B3C527B3D} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1835D3A2-5ED1-4261-A502-4AE056427BC7}
