+++ src/AspireApp.ServiceDefaults/ServiceDefaultsExtensions.cs
@@ -99,7 +99,8 @@ public static class ServiceDefaultsExtensions
 
     public static WebApplication MapDefaultEndpoints(this WebApplication app)
     {
-        if (app is null) throw new ArgumentNullException(nameof(app));
+        ArgumentNullException.ThrowIfNull(app);
+
         // Adding health checks endpoints to applications in non-development environments has security implications.
         // See https://aka.ms/dotnet/aspire/healthchecks for details before enabling these endpoints in non-development environments.
         if (app.Environment.IsDevelopment())
