+++ src/AspireApp.Shared/DTOs/UpdateProductDto.cs
@@ -1,0 +1,10 @@
+namespace AspireApp.SharedLib.DTOs;
+
+/// <summary>
+/// DTO for updating an existing product
+/// </summary>
+public record UpdateProductDto(
+    string? Name,
+    string? Description,
+    decimal? Price,
+    int? Stock);
